Estimate the PSY statistic sequence

PSY implements the real time bubble detection procedure of Phillips, Shi and Yu (2015a,b)

PSY(y, swindow0, IC = 0, adflag = 0)

Arguments

y
A vector. The data.

swindow0
A positive integer. Minimum window size.

IC
A positive integer. 0 for fixed lag order 1 for AIC and 2 for BIC (default = 0).

adflag
A positive integer, lag order when IC=0; maximum number of lags when IC>0 (default = 0).

Value

Vector, BSADF test statistic.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015a). Testing for multiple bubbles: Historical episodes of exuberance and collapse in the S&P 500. International Economic Review, 56(4), 1034--1078.

Phillips, P. C. B., Shi, S., & Yu, J. (2015b). Testing for multiple bubbles: Limit Theory for Real-Time Detectors. International Economic Review, 56(4), 1079--1134.

Examples

y     <- rnorm(100)
BSADF <- PSY(y,  swindow0 = 90, IC = 0, adflag = 1)